Switch from pygettext to xgettext (#248)
* Switch from pygettext to xgettext

This commit switches the update_translations.py script to use
xgettext instead of pygettext, mostly because xgettext is installed
system wide, while pygettext moves around, especially when updating
python versions.

It also contains a few tweaks to the Naomi TTI system, which was
scoring Hacker News higher than News when responding to the word
"news" because every template in Hacker News contains the word
news while one of the templates in News does not. I have therefore
started penalizing the intent proportional to the weight of words
that do not appear in the incoming request, which seems to work
well so far, and penalizes "Hacker News" into a lower ranking if
the word "Hacker" is not detected in the incoming request.

* Skip WER if transcription is blank

jiWER suddenly is throwing an error if the transcription is
blank, so skip it.

* Fixes to update_translations.py

Fixed a couple of bugs introduced when updating update_translations.py
to python3, and added some fixes to hopefully prevent msgcat from
duplicating the header information when updating translations.

* Flake8 issues

Fixed some formatting issues, unused variables, unused import

* Codacy subprocess complaint

Codacy does not like subprocess. Added a nosec check.

* Fixed plurals

Fixed a bug which was preventing the plurals information from
being written to the .po files

* Better recovery after canceling setup

I started setting up Naomi, but canceled the process before it
completed. At that point, it had built the Pocketsphinx model,
but had not set the locations of the hmm dir or fst file in the
profile. Because it did not have to create them the next time I
ran it, it did not set the default location of the fst file.

* STTTrainer fix for new intents

This fixes the Adapt Pocketsphinx STTTrainer plugin to extract
intent templates from the new intent structure.

+++ plugins/stt/pocketsphinx-stt/sphinxplugin.py
@@ -66,6 +66,8 @@ def check_pocketsphinx_model(directory):
         FilesExist = False
     if(not os.path.isfile(os.path.join(directory, "variances"))):
         FilesExist = False
+    if(not os.path.isfile(os.path.join(directory, "model", "train.fst"))):
+        FilesExist = False
     return FilesExist
 
 
@@ -159,6 +161,7 @@ class PocketsphinxSTTPlugin(plugin.STTPlugin):
             os.remove(self._logfile)
 
     def settings(self):
+        language = profile.get(['language'])
         # Get the defaults for settings
         # hmm_dir
         hmm_dir = profile.get(
@@ -254,7 +257,7 @@ class PocketsphinxSTTPlugin(plugin.STTPlugin):
                     fst_model = path
         # If either the hmm dir or fst model is missing, then
         # download the standard model
-        if not(hmm_dir and fst_model):
+        if not(hmm_dir and os.path.isdir(hmm_dir) and fst_model and os.path.isfile(fst_model)):
             # Start by checking to see if we have a copy of the standard
             # model for this user's chosen language and download it if not.
             # Check for the files we need
@@ -269,6 +272,7 @@ class PocketsphinxSTTPlugin(plugin.STTPlugin):
             if not os.path.isdir(standard_dir):
                 os.mkdir(standard_dir)
             hmm_dir = standard_dir
+            fst_model = os.path.join(hmm_dir, "train", "model.fst")
             formatteddict_path = os.path.join(
                 hmm_dir,
                 "cmudict.formatted.dict"
@@ -287,7 +291,7 @@ class PocketsphinxSTTPlugin(plugin.STTPlugin):
                 ]
                 completedprocess = run_command(cmd)
                 self._logger.info(process_completedprocess(completedprocess))
-
+            if(not os.path.isfile(formatteddict_path)):
                 print("Formatting the g2p dictionary")
                 with open(os.path.join(standard_dir, "cmudict.dict"), "r") as in_file:
                     with open(formatteddict_path, "w+") as out_file:
@@ -301,6 +305,7 @@ class PocketsphinxSTTPlugin(plugin.STTPlugin):
                             # replace the first whitespace with a tab
                             line = line.replace(' ', '\t', 1)
                             print(line, file=out_file)
+            if(not os.path.isfile(fst_model)):
                 # Use phonetisaurus to prepare an fst model
                 print("Training an FST model")
                 cmd = [
@@ -311,7 +316,6 @@ class PocketsphinxSTTPlugin(plugin.STTPlugin):
                 ]
                 completedprocess = run_command(cmd)
                 self._logger.info(process_completedprocess(completedprocess))
-                fst_model = os.path.join(hmm_dir, "train", "model.fst")
 
         phonetisaurus_executable = profile.get_profile_var(
             ['pocketsphinx', 'phonetisaurus_executable']
